requirements.txt
absl-py==0.12.0
astunparse==1.6.3
cached-property==1.5.2
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
flatbuffers==1.12
gast==0.3.3
google-auth==1.27.1
google-auth-oauthlib==0.4.3
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
importlib-metadata==3.7.2
Keras~=2.12.0
Keras-Preprocessing==1.1.2
Markdown==3.3.4
numpy~=1.23.5
oauthlib==3.1.0
opencv-python~=4.7.0.72
opt-einsum==3.3.0
Pillow~=9.4.0
protobuf==3.15.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.5.4
six==1.15.0
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow~=2.12.0
tensorflow-estimator==2.4.0
termcolor==1.1.0
typing-extensions==3.7.4.3
urllib3==1.26.3
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.1

pygame~=2.3.0
mysql-connector-python~=8.0.32
pydub~=0.25.1
mutagen~=1.46.0
matplotlib~=3.7.1
scikit-learn~=1.2.2